Refactor internal.Cluster type
Remove any AWS specific informations from Cluster type and move it
to a separate file. We can use the same type for all cloud providers
